bugfix: cd back to where we started
The `mkdir`/`cp` commands weren't behaving correctly because we were in
the wrong directory

I think the *right* fix is probably to never `cd` at all? But that
requires a pretty strong understanding of how all the sed/make/whatever
commands we're using handle directories. I do not currently have that
understanding.